Move commission overflow test into peersUpdate_test.go
