+++ core/kapp/validators/peersUpdate_test.go
@@ -2205,3 +2205,67 @@ func TestProcessEconomicsEndOfEpoch_V1V2(t *testing.T) {
 	})
 
 }
+
+func TestV2CommissionBigIntHandlesOverflow(t *testing.T) {
+	rewardsAddr := makeAddress("rewards-address")
+	delegatorAddr := makeAddress("delegator-1")
+	const delegatedAmount int64 = 1_000_000
+
+	run := func(t *testing.T, accumulatedFees int64, commission uint32) map[string]int64 {
+		require.Less(t, accumulatedFees*int64(commission), int64(0),
+			"scenario must be in the int64 overflow regime so it exercises the fix")
+
+		args := createMockArgs()
+		v, err := NewValidatorKApp(args)
+		require.NoError(t, err)
+		v.forkController.(*mock.ForkControllerStub).EnableSmartContractsValue = true
+
+		val := &ValidatorData{
+			RewardsAddress: rewardsAddr,
+			Commission:     commission,
+		}
+		accumulatedDelegations := map[string]int64{
+			string(delegatorAddr): delegatedAmount,
+		}
+
+		return v.calculateDelegationRewardsV2(val, accumulatedFees, accumulatedDelegations, delegatedAmount)
+	}
+
+	sumOf := func(m map[string]int64) int64 {
+		var sum int64
+		for _, amt := range m {
+			sum += amt
+		}
+		return sum
+	}
+
+	t.Run("FullRateCommission", func(t *testing.T) {
+		accumulatedFees := int64(1_000_000_000_000_000)
+		local := run(t, accumulatedFees, uint32(core.HundredPercent))
+
+		require.Equal(t, accumulatedFees, local[string(rewardsAddr)],
+			"full-rate commission must equal accumulatedFees with big.Int, not overflow negative")
+		require.GreaterOrEqual(t, local[string(rewardsAddr)], int64(0),
+			"commission must not be negative")
+		require.Equal(t, accumulatedFees, sumOf(local),
+			"distribution must telescope to accumulatedFees")
+	})
+
+	t.Run("PartialRateExercisesDistribution", func(t *testing.T) {
+		accumulatedFees := int64(2_000_000_000_000_000)
+		commission := uint32(core.HundredPercent / 2)
+		local := run(t, accumulatedFees, commission)
+
+		commissionEntry := local[string(rewardsAddr)]
+		delegatorEntry := local[string(delegatorAddr)]
+
+		require.GreaterOrEqual(t, commissionEntry, int64(0),
+			"commission must not be negative")
+		require.Greater(t, delegatorEntry, int64(0),
+			"delegator share must be non-zero so the distribution path is genuinely exercised")
+		require.LessOrEqual(t, commissionEntry, accumulatedFees,
+			"commission must not exceed accumulatedFees")
+		require.Equal(t, accumulatedFees, sumOf(local),
+			"distribution must telescope to accumulatedFees through the delegator-share path")
+	})
+}
